[REF] Article, Comment Controller Test에 예외관련 Case들 추가

+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -60,8 +60,15 @@ public class ArticleService {
     @Transactional
     public void deleteById(long id) {
         log.debug("delete article id={}", id);
+        if (isArticleNotFound(id)) {
+            throw new ArticleNotFoundException(ERROR_ARTICLE_NOT_FOUND_MESSAGE);
+        }
 
         commentRepository.deleteByArticleId(id);
         articleRepository.deleteById(id);
+    }
+
+    private boolean isArticleNotFound(long id) {
+        return findById(id) == null;
     }
 }
